requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Limiter
Flask-CORS
python-dotenv
email-validator
psycopg2